Ignore `StaleElementReferenceException` in `_sign_in_success`
The sign in page sends the authentication request asynchronously before
the rest of the page navigates away. The `user-greeting` element may be
present when `find_element` runs, but could then be gone by the time the
`text` attribute is (attempted to be) accessed. If the element is gone
then sign in is not (yet) successful, so this exception can be ignored.